# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SysMap Social Network - OpenAPI 3.0",
    "description": "Backend of the Social Network project of SysMap fullstack bootcamp.",
    "contact": {
      "email": "kellyplcastelo@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/v1"
    }
  ],
...
  "paths": {
...
    "/posts/{id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Find posts by id",
        "description": "Find posts by id",
        "operationId": "findPostById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to get"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "put": {
        "tags": [
          "posts"
        ],
        "summary": "Update a post",
        "description": "This can only be done by the logged in user.",
        "operationId": "updatePost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to update"
          }
        ],
        "requestBody": {
          "description": "Update an existent post",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostRequest"
              }
            }
          }
        },
        "responses": {
          "203": {
            "description": "UNDOCUMENTED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "Deletes a post",
        "description": "delete a post",
        "operationId": "deletePost",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "NO CONTENT"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
...
    "/posts/{postId}/comments/{id}": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "Find comments by id",
        "description": "Find comments by id",
        "operationId": "findCommentsById",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to get the comment from"
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the comment"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "put": {
        "tags": [
          "comments"
        ],
        "summary": "Update a comment",
        "description": "Update a comment",
        "operationId": "updateComment",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to get the comment from"
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the comment"
          }
        ],
        "requestBody": {
          "description": "Update an existent post",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "203": {
            "description": "UNDOCUMENTED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      },
      "delete": {
        "tags": [
          "comments"
        ],
        "summary": "Deletes a comment from a post",
        "description": "Deletes a comment from a post",
        "operationId": "deleteComment",
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the post to get the comment from"
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the comment"
          }
        ],
        "responses": {
          "204": {
            "description": "NO CONTENT"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
...
    "/profile/:id/follow": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Follow a profile",
        "description": "Follow a profile, returns 400 if you're already following this profile",
        "operationId": "followProfile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "uuid of the profile to follow"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST"
          },
          "401": {
            "description": "UNAUTHORIZED"
          },
          "404": {
            "description": "NOT FOUND"
          }
        },
        "security": [
          {
            "BearerAuth": {}
          }
        ]
      }
    },
...
    "/security/register": {
      "post": {
        "tags": [
          "security"
        ],
        "summary": "Creates a new user",
        "description": "Creates a new user",
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CREATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorArrayResponse"
                }
              }
            }
          }
        }
      }
    },
    "/security/login": {
      "post": {
        "tags": [
          "security"
        ],
        "summary": "Logs in an user",
        "description": "Logs in an user",
        "operationId": "logInUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "minLength": 2
          },
          "password": {
            "type": "string",
            "minLength": 2
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJfaWQiOiI2MzdmM2E2ZTJjODJjNjYwYTk2ZDhjY2YiLCJ1c2VyIjoiYmFyYmEiLCJfX3YiOjAsInByb2ZpbGUiOnsiX2lkIjoiNjM3ZjNhNmUyYzgyYzY2MGE5NmQ4Y2Q1IiwibmFtZSI6ImJhcmJhIiwidXNlciI6IjYzN2YzYTZlMmM4MmM2NjBhOTZkOGNjZiIsImZvbGxvd2luZyI6WyI2MzdmM2E2ZTJjODJjNjYwYTk2ZDhjZTgiLCI2MzdmM2E2ZTJjODJjNjYwYTk2ZDhjZDUiXSwiZm9sbG93ZXJzIjpbIjYzN2YzYTZlMmM4MmM2NjBhOTZkOGNkMSIsIjYzN2YzYTZlMmM4MmM2NjBhOTZkOGNkNSJdLCJfX3YiOjB9fQ.0KzLRscYliEFmUl61KWE5MEKS69XJWfalL5hfpy-BG8"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "example": "Kelly"
          },
          "user": {
            "type": "string",
            "minLength": 2,
            "example": "Castelo"
          },
          "password": {
            "type": "string",
            "minLength": 2,
            "example": "randompassoword123"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637c9bb1fceec2bc7d8d9249"
          },
          "user": {
            "type": "string",
            "example": "Castelo"
          }
        }
      },
      "ErrorArrayResponse": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "/v1/path/action"
          },
          "status": {
            "type": "integer",
            "example": "400"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "Path `user` is required"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "/v1/path/action"
          },
          "status": {
            "type": "integer"
          },
          "errors": {
            "type": "string"
          }
        }
      },
      "PostResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bbc3a69c083c2b0003097"
          },
          "title": {
            "type": "string",
            "example": "To be or not to be"
          },
          "description": {
            "type": "string",
            "example": "That is the question"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          }
        }
      },
...
      "PostRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "To be or not to be"
          },
          "description": {
            "type": "string",
            "example": "That is the question"
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7c1"
          },
          "name": {
            "type": "string",
            "example": "Kelly Castelo"
          },
          "user": {
            "type": "string",
            "example": "637bb9d4da4e28ea8606b7bf"
          },
          "following": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c7"
            }
          },
          "followers": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c8"
            }
          }
        }
      },
...
      "CommentRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Amazing post!"
          }
        },
        "required": [
          "description"
        ]
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "637bbc3a69c083c2b0003097"
          },
          "description": {
            "type": "string",
            "example": "That is the question"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "637bb9d4da4e28ea8606b7c4"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}